Move builtin settings data file to correct directory
